Extract TestSceneInputQueueChange
